refactor: comprehensive module updates and test cleanup
Major improvements across core trading system modules with extensive
test consolidation and cleanup.

## Core Module Updates

### Bot Management
- Enhanced bot entity lifecycle management and state transitions
- Improved resource allocation and coordination services
- Updated bot instance and lifecycle service implementations
- Refined coordination service for multi-bot operations

### Capital Management
- Updated DI registration for better dependency management
- Enhanced exchange distributor and capital allocation logic
- Improved factory patterns and service implementations

### Risk Management
- Enhanced risk controller with better validation
- Improved adaptive risk management and stop-loss logic
- Updated position sizing calculations (Kelly, fixed, volatility)
- Better risk manager unit implementations

### Strategies
- Updated all strategy implementations (mean reversion, trend following, etc.)
- Enhanced strategy factory and DI registration
- Improved market making and arbitrage strategies
- Better strategy validation and performance monitoring

### Exchanges
- Updated Coinbase, OKX, and Mock exchange implementations
- Enhanced exchange DI registration
- Improved exchange error handling and validation

### Data & Monitoring
- Enhanced data quality validation and monitoring
- Improved telemetry and performance tracking
- Updated metrics collection and monitoring services

### Execution
- Enhanced execution engine and orchestration
- Improved execution service implementations

### ML & Analytics
- Updated ML service and feature engineering
- Enhanced model registry
- Improved analytics DI registration

## Test Infrastructure

### Test Cleanup (Deleted 100+ obsolete files)
- Removed duplicate integration tests across all modules
- Deleted legacy validation tests replaced by newer implementations
- Consolidated overlapping test scenarios
- Removed disabled/backup test files

### Test Updates
- Updated unit tests across all modules
- Enhanced integration test coverage
- Improved test fixtures and conftest configurations
- Better test isolation and cleanup

### Notable Test Improvements
- Bot management: Consolidated from 7 to 1 real integration test
- Database: Removed 10 duplicate test files
- Strategies: Updated all strategy tests with better coverage
- Risk management: Enhanced adaptive and validation tests

## Documentation
- Updated REFERENCE.md files across all modules
- Better API documentation and usage examples

## Statistics
- 302 files changed
- ~20,500 insertions, ~61,800 deletions (net reduction through cleanup)
- Removed 100+ duplicate/legacy test files
- Updated 200+ source and test files

+++ src/core/types/strategy.py
@@ -86,28 +86,28 @@ class StrategyConfig(BaseModel):
 
     # Common strategy parameters
     lookback_period: int | None = None
-    entry_threshold: float | None = None
-    exit_threshold: float | None = None
-    stop_loss_pct: float | None = None
-    take_profit_pct: float | None = None
+    entry_threshold: Decimal | None = None
+    exit_threshold: Decimal | None = None
+    stop_loss_pct: Decimal | None = None
+    take_profit_pct: Decimal | None = None
     max_position_size: Decimal | None = None
-    min_confidence: float | None = None
+    min_confidence: Decimal | None = None
     max_positions: int | None = None
-    position_size_pct: float | None = None
+    position_size_pct: Decimal | None = None
     exchange_type: str | None = None
     requires_risk_manager: bool = False
     requires_exchange: bool = False
 
     # Mean reversion specific
     window_size: int | None = None
-    num_std: float | None = None
+    num_std: Decimal | None = None
 
     # Momentum specific
-    momentum_threshold: float | None = None
-    acceleration_factor: float | None = None
+    momentum_threshold: Decimal | None = None
+    acceleration_factor: Decimal | None = None
 
     # Market making specific
-    spread_pct: float | None = None
+    spread_pct: Decimal | None = None
     inventory_target: Decimal | None = None
 
     created_at: datetime = Field(default_factory=lambda: datetime.now(timezone.utc))
